requirements.txt
astroid==2.3.0
awscli==1.18.172
backcall==0.2.0
botocore==1.19.12
colorama==0.4.3
decorator==4.4.2
Django==2.0.2
docutils==0.15.2
git-remote-codecommit==1.15.1
ikp3db==1.4.1
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.11.1
jmespath==0.10.0
lazy-object-proxy==1.5.1
mccabe==0.6.1
parso==0.1.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.7.2
pylint==2.4.4
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.15.0
stevedore==3.2.2
traitlets==4.3.3
typed-ast==1.2.0
urllib3==1.25.11
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
